db-virtualization: Renamed db.py to base_model.py
