+++ webserver/app/__init__.py
@@ -15,7 +15,7 @@ from app import (
     main, admin_api, datasets_api, tasks_api, requests_api,
     containers_api, registries_api, users_api
 )
-from app.helpers.db import build_sql_uri, db
+from webserver.app.helpers.base_model import build_sql_uri, db
 from app.helpers.exceptions import (
     InvalidDBEntry, DBError, DBRecordNotFoundError, InvalidRequest,
     AuthenticationError, UnauthorizedError, KeycloakError, TaskImageException,
