+++ webserver/app/datasets_api.py
@@ -15,7 +15,7 @@ from datetime import datetime
 from flask import Blueprint, request
 
 from .helpers.exceptions import DBRecordNotFoundError, InvalidRequest
-from .helpers.db import db
+from .helpers.base_model import db
 from .helpers.keycloak import Keycloak
 from .helpers.query_validator import validate
 from .helpers.wrappers import auth, audit
