+++ webserver/app/models/dictionary.py
@@ -2,7 +2,7 @@ from datetime import datetime
 from sqlalchemy import Column, Integer, DateTime, String, ForeignKey, UniqueConstraint
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
-from app.helpers.db import BaseModel, db
+from webserver.app.helpers.base_model import BaseModel, db
 from app.helpers.exceptions import InvalidRequest
 from app.models.dataset import Dataset
