+++ webserver/app/models/request.py
@@ -4,7 +4,7 @@ from sqlalchemy import Column, Integer, DateTime, String, ForeignKey, update
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 from sqlalchemy.exc import IntegrityError
-from app.helpers.db import BaseModel, db
+from webserver.app.helpers.base_model import BaseModel, db
 from app.models.dataset import Dataset
 from app.helpers.keycloak import Keycloak
 from app.helpers.exceptions import DBError
