+++ webserver/tests/test_datasets.py
@@ -8,7 +8,7 @@ from sqlalchemy.exc import ProgrammingError, OperationalError
 from unittest import mock
 from unittest.mock import Mock
 
-from app.helpers.db import db
+from webserver.app.helpers.base_model import db
 from app.helpers.exceptions import KeycloakError
 from app.models.dataset import Dataset
 from app.models.catalogue import Catalogue
